requirements.txt
annotated-types==0.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
docstring_parser==0.16
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-aiplatform==1.52.0
google-cloud-bigquery==3.23.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.64.0
grpcio-status==1.62.2
idna==3.7
numpy==1.26.4
packaging==24.0
pandas==2.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pypdf==4.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.2
rsa==4.9
shapely==2.0.4
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
